Refactored Settings views
