+++ Tuneful/Tuneful.swift
@@ -93,11 +93,9 @@ class AppDelegate: NSObject, NSApplicationDelegate, NSMenuDelegate {
             object: nil
         )
         
-//        if !viewedShortcutsSetup {
+        if !viewedShortcutsSetup {
             self.showShortcutsSetup()
-//        }
-        
-        self.showOnboarding()
+        }
         
         if !viewedOnboarding {
             self.showOnboarding()
